requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Kemp\Desktop\JupyterDump\water-bot\water_bot\water_bot.py: 3
discord.py == 1.3.3

# C:\Users\Kemp\Desktop\JupyterDump\water-bot\setup.py: 5
setuptools == 46.1.3.post20200330
